+++ libraries/api-extractor/src/Token.ts
@@ -25,7 +25,7 @@ export enum TokenType {
 }
 
 /**
- * A structured object created from a doc comment string within a JSDoc comment block.
+ * A structured object created from a doc comment string within an AEDoc comment block.
  */
 export default class Token {