+++ libraries/api-extractor/src/generators/test/ApiFileGenerator.test.ts
@@ -63,8 +63,8 @@ describe('ApiFileGenerator tests', function (): void {
        * Errors can be found in testInputs/folder/MyClass
        */
       assert.equal(capturedErrors.length, 2);
-      assert.equal(capturedErrors[0].message, 'Unknown JSDoc tag "@badjsdoctag"');
-      assert.equal(capturedErrors[1].message, 'Unexpected text in JSDoc comment: '
+      assert.equal(capturedErrors[0].message, 'Unknown AEDoc tag "@badAedocTag"');
+      assert.equal(capturedErrors[1].message, 'Unexpected text in AEDoc comment: '
         + '"(Error #1 is the bad tag) Text can no..."');
     });