+++ libraries/api-extractor/testInputs/example1/folder/MyClass.ts
@@ -1,7 +1,7 @@
 
 /**
-  * @badjsdoctag (Error #1 is the bad tag) Text can not come after a tag unless it is a parameter of
-  * the tag. It must come in the first few sentences of the JSDoc or come after 
+  * @badAedocTag (Error #1 is the bad tag) Text can not come after a tag unless it is a parameter of
+  * the tag. It must come in the first few sentences of the AEDoc or come after
   * an \@internalremarks tag. (Error #2 text coming after a tag that is not \@internalremarks)
   */
 export default class MyClass {
@@ -32,8 +32,8 @@ class PrivateClass {
  */
 export class InternalClass {
   /**
-   * This will not report an error, instead a 'WARNING' comment will 
-   * appear above this class declaration in the API file as a result of 
+   * This will not report an error, instead a 'WARNING' comment will
+   * appear above this class declaration in the API file as a result of
    * no type declaration.
    */
   public static propertyWithNoType;
@@ -72,7 +72,7 @@ export class PreapprovedInternalClass {
 const privateField = 123;
 
 /**
- * This is testing identifiers whose name is the same as the 
+ * This is testing identifiers whose name is the same as the
  * members of the object prototype.
  */
 export class __proto__ {
@@ -86,6 +86,6 @@ export interface hasOwnProperty {
 export class A extends __proto__ implements hasOwnProperty {
     __lookupSetter__: __proto__;
     public __proto__(__proto__: string): __proto__ {
-        return undefined;   
+        return undefined;
     }
 }